Move HiveClassNames to hive-formats module
